+++ src/MachO/json_internal.hpp
@@ -46,6 +46,7 @@ class Header;
 class LinkerOptHint;
 class LoadCommand;
 class MainCommand;
+class Routine;
 class RPathCommand;
 class Relocation;
 class RelocationDyld;
@@ -94,6 +95,7 @@ class JsonVisitor : public LIEF::JsonVisitor {
   void visit(const LinkerOptHint& e)                      override;
   void visit(const LoadCommand& cmd)                      override;
   void visit(const MainCommand& maincmd)                  override;
+  void visit(const Routine& rpath)                        override;
   void visit(const RPathCommand& rpath)                   override;
   void visit(const Relocation& relocation)                override;
   void visit(const RelocationDyld& rdyld)                 override;