+++ doc/sphinx/changelog.rst
@@ -4,6 +4,11 @@ Changelog
 0.16.0 - Not Released Yet
 -------------------------
 
+:MachO:
+
+  * Add support for the ``LC_SUBCLIENT`` command: :class:`lief.MachO.SubClient`
+  * Add support for the ``LC_ROUTINE/LC_ROUTINE64`` command: :class:`lief.MachO.Routine`
+
 
 0.15.1 - July 23th, 2024
 ------------------------